Fix: fix mixing in pair tasks (track-track & track-v0) (#5389)
It seems that there is a bug in the central mixing framework when using
partition of tables. Go back to the original implementation, where the
underlying table of a partition is passed (which may results in larger
memory usage, but the mixing is working as intended) to the combination
object.