PWGCF : Femtouniverse -- Fixing bug in dphi calculation (#5643)
* PWGCF : Femtouniverse -- Fixing bug in dphi calculation

The extra factor of 0.1 that was scaling the magnetic field is removed. 
However, after removing this factor, the asin(<arg>) function for calculating dphi was blowing up at certain (mostly higher) values of radii. This is also taken care of by only allowing abs(<arg>) < 1.0.

* PWGCF : Femtouniverse -- Fixing bug in dphi calculation

The extra factor of 0.1 that was scaling the magnetic field is removed.
However, after removing this factor, the asin() function for calculating dphi was blowing up at certain (mostly higher) values of radii. This is also taken care of by only allowing abs() < 1.0.

* Update FemtoUniverseDetaDphiStar.h

* Update FemtoUniverseDetaDphiStar.h